Refactors reducer and adds sombineReducers from redux
